backend/pyproject.toml
[project]
name = "seedling-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.135.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
    "boto3>=1.42.68",
    "python-multipart>=0.0.22",
    "pillow>=12.1.1",
    "pydantic-settings>=2.13.1",
    "pytest>=9.0.2",
    "behave>=1.3.3",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
]
